refactor: use decorator for cache read/write
`League` methods decorated with `@cache_operator` will have their
non-self params passed to the `load` and `save` methods of the instance's
cache attribute. If the load succeeds, the data are returned. Otherwise the
decorated function is called, and the returned data is saved to the cache.
This decorator relies on the cache API remaining consistent, which would
likely not be the case if other ESPN endpoints were to be used in the future.